feat: implemented the error with snackbar

+++ app/src/main/java/com/github/se/assocify/ui/screens/profile/treasuryTags/ProfileTreasuryTagsViewModel.kt
@@ -1,10 +1,12 @@
 package com.github.se.assocify.ui.screens.profile.treasuryTags
 
+import androidx.compose.material3.SnackbarHostState
 import androidx.lifecycle.ViewModel
 import com.github.se.assocify.model.CurrentUser
 import com.github.se.assocify.model.database.AccountingCategoryAPI
 import com.github.se.assocify.model.entities.AccountingCategory
 import com.github.se.assocify.navigation.NavigationActions
+import com.github.se.assocify.ui.util.SnackbarSystem
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
@@ -14,6 +16,8 @@ class ProfileTreasuryTagsViewModel(
 ) : ViewModel() {
   private val _uiState = MutableStateFlow(ProfileTreasuryTagsUIState())
   val uiState: StateFlow<ProfileTreasuryTagsUIState> = _uiState
+
+  private val snackBarSystem = SnackbarSystem(_uiState.value.snackBarHostState)
 
   init {
     fetchCategories()
@@ -41,26 +45,37 @@ class ProfileTreasuryTagsViewModel(
   }
 
   fun deleteTag(tag: AccountingCategory) {
+    if (_uiState.value.nameError) return
     accountingCategoryAPI.deleteCategory(
         tag,
         {
           _uiState.value =
               _uiState.value.copy(treasuryTags = _uiState.value.treasuryTags.filter { tag != it })
         },
-        {})
+        {
+          snackBarSystem.showSnackbar(
+              "Could not delete the tag",
+          )
+        })
   }
 
   fun addTag(newTag: AccountingCategory) {
+    if (_uiState.value.nameError) return
     accountingCategoryAPI.addCategory(
         associationUID = _uiState.value.assocId,
         newTag,
         {
           _uiState.value = _uiState.value.copy(treasuryTags = _uiState.value.treasuryTags + newTag)
         },
-        {})
+        {
+          snackBarSystem.showSnackbar(
+              "Could not add the tag",
+          )
+        })
   }
 
   fun modifyTag(tag: AccountingCategory) {
+    if (_uiState.value.nameError) return
     accountingCategoryAPI.updateCategory(
         _uiState.value.assocId,
         tag,
@@ -69,12 +84,22 @@ class ProfileTreasuryTagsViewModel(
               _uiState.value.copy(
                   treasuryTags = _uiState.value.treasuryTags.filter { it.uid != tag.uid } + tag)
         },
-        {})
+        {
+          snackBarSystem.showSnackbar(
+              "Could not modify the tag",
+          )
+        })
   }
 
   fun cancelPopUp() {
     _uiState.value =
         _uiState.value.copy(modify = false, creating = false, editedTag = null, displayedName = "")
+  }
+
+  fun checkNameError(name: String) {
+    _uiState.value =
+        _uiState.value.copy(
+            nameError = name.isEmpty() || _uiState.value.treasuryTags.any { ac -> ac.name == name })
   }
 }
 
@@ -86,5 +111,8 @@ data class ProfileTreasuryTagsUIState(
     val editedTag: AccountingCategory? = null,
     val displayedName: String = "",
     val error: String? = null,
-    val loading: Boolean = false
+    val loading: Boolean = false,
+    val snackBarError: String? = null,
+    val nameError: Boolean = false,
+    val snackBarHostState: SnackbarHostState = SnackbarHostState()
 )
